% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_NA.R
\name{cbind.NA}
\alias{cbind.NA}
\title{cbind data frames of varying lengths}
\usage{
\method{cbind}{`NA`}(..., fill = NA)
}
\arguments{
\item{...}{}

\item{fill}{}
}
\value{
data.frame
}
\description{
cbind data frames of varying lengths
}
