% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCombine.R
\name{combCores}
\alias{combCores}
\title{Combine multiple ring width measurements from the same sample}
\usage{
combCores(rwi, PO = NULL)
}
\arguments{
\item{rwi}{}
}
\value{
rwi
}
\description{
Combine multiple ring width measurements from the same sample
}
