% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRCS.R
\name{dRCS}
\alias{dRCS}
\title{regional curve standardization with plotting and custom RC input}
\usage{
dRCS(rwl, po, rc)
}
\arguments{
\item{rwl}{rwl file as from dplR}

\item{po}{po file as from dplR}

\item{rc}{regional curve formatted as the RC object returned from robustRC()}
}
\value{
series info, rwi, plot data, plot
}
\description{
regional curve standardization with plotting and custom RC input
}
