% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenOut.R
\name{evenOut}
\alias{evenOut}
\title{look for inconsistency in curve fit by age}
\usage{
evenOut(AgeAlin)
}
\arguments{
\item{AgeAlin}{age aligned ring widths}
}
\value{
adjusted age aligned
}
\description{
There should probably be a citation here, but I didn't write it down when I first wrote this code
}
