% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRCS1.R
\name{newRCS}
\alias{newRCS}
\title{RCS (regional curve standardization) detrending}
\usage{
newRCS(
  rwlFile = NULL,
  poFile = NULL,
  ratios = TRUE,
  truncRC = 10,
  rcIn = NULL,
  ageMin = NULL,
  ageMax = NULL
)
}
\arguments{
\item{rwlFile}{as created by read.rwl}

\item{poFile}{two-column file created by dplR}

\item{ratios}{use ratios rather than differences (recommended)}

\item{truncRC}{limit curve calculation to regions of a designated minimum sample depth}

\item{rcIn}{two column regional curve as created from robustRC() or NULL to create new}

\item{ageMin}{max age of valid RC, or NULL if no truncation needed}

\item{ageMax}{min age of valid RC, or NULL if no truncation needed}
}
\value{
rwi and assorted goodies
}
\description{
RCS (regional curve standardization) detrending
}
