% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTrunc.R
\name{ontoTrunc}
\alias{ontoTrunc}
\title{truncate the series in the rwl file with min/max ontogenetic ages}
\usage{
ontoTrunc(rwl, po, ontoCut)
}
\arguments{
\item{rwl}{as produced by dplR}

\item{po}{as produced by dplR}

\item{ontoCut}{c(x,y)}
}
\value{
rwl, po
}
\description{
truncate the series in the rwl file with min/max ontogenetic ages
}
