% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustRC.R
\name{robustRC}
\alias{robustRC}
\title{create a regional curve (average ring with by ontogenetic age)}
\usage{
robustRC(
  rwlFile = NULL,
  poFile = NULL,
  truncRC = 20,
  aAligned = FALSE,
  evO = FALSE,
  dMethod = NULL,
  spLen = NULL,
  tvSpline = TRUE,
  tvRange = c(3, 80),
  tvStiff = NULL
)
}
\arguments{
\item{rwlFile}{as created by read.rwl in dplR}

\item{poFile}{two-column file created by dplR}

\item{truncRC}{limit curve calculation to regions of a designated minimum sample depth}

\item{aAligned}{TRUE/FALSE, is the rwl file already aligned by ontogenetic age?}

\item{evO}{adjust RC fit (see evenOut())}

\item{dMethod}{how should we fit the curve to the ontogenetic age values? methods from dplR::detrend()}

\item{spLen}{if dMethod is "Spline", what spline length (in number of years)}

\item{tvSpline}{TRUE/FALSE, use an age-varying spline (recommended, dMethod must be set to NULL)}

\item{tvRange}{min and max spline lengths in nyrs (if using tvSpline), eg. c(3,80)}

\item{tvStiff}{set the age at which the spline length will max out (if using tvSpline)}
}
\value{
curve, info, and plot
}
\description{
create a regional curve (average ring with by ontogenetic age)
}
