% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvSpline.R
\name{tvSpline}
\alias{tvSpline}
\title{Age-varying spline}
\usage{
tvSpline(series, SpLenRange = c(3, 100), goStiff = NULL, plot = FALSE)
}
\arguments{
\item{series}{time series}

\item{SpLenRange}{min/max spline length in nyrs eg (3,80)}

\item{goStiff}{at what age should the spline length max out? or NULL}
}
\value{
original series and spline
}
\description{
Age-varying spline
}
