% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCombine.R
\name{combCores}
\alias{combCores}
\title{Combine multiple ring width measurements from the same sample}
\usage{
combCores(rwi, po)
}
\arguments{
\item{rwi}{ring width data as data.frame from dplR}

\item{po}{pith offset data with columns for sample ID and pith offset}
}
\value{
list with combined rwl and updated PO
}
\description{
Combine multiple ring width measurements from the same sample
}
